RAGelectronics/requirements.txt
streamlit==1.33.0
PyPDF2==3.0.1
langchain==0.1.1
langchain-text-splitters==0.0.1
langchain-community==0.0.15
langchain-huggingface==0.0.1
pinecone-client==3.0.0
pinecone-text==0.7.0
python-dotenv==1.0.0
nltk==3.8.1
pandas==2.1.3
SQLAlchemy==2.0.23